Remove fields for brevity in JSON
The "date" and "account" fields are removed in the JSON list. Currently
urls seem to be limited to a certain number of characters (deduced from
trial and error). To allow more entries to be incorporated in a single url,
remove the above two fields which I think are less important.

The changes also remove some debug prints statements I accidentally left in.