Send <=3 transactions; Don't send null transaction
There seems to be a character limit on the Cashew deep link; if longer,
the url cannot be tapped to fill entries. Send at most three transactions
per message, and send multiple messages if more transactions exist.

Additionally, fix a bug where a `None` value was being appended to the
transaction array. This is from emails from a bank, but which does not
contain transaction information.