Use email date for Cashew entries; Url encode periods
The issue with iMessage not interpreting a Cashew deep link as something
clickable was because of presence of some periods (this doesn't occur every
time the link contains a period). Url encode periods to avoid this issue.

Since the length is not a problem, we can add dates to the entries. Use the
date of the email as the date of the transaction.